Re-added the bit more detailed logging output of the input sentence.
Our bimachines would now be perfect, if we used regular a
(character-based) alphabet. But with these symbols, we are a bit
unlucky: the left-right sequential transduction loses the information
encoded in the identity Symbols, i.e. the actual text token associated
with the symbol. Let's think a bit how we could preserve this information...